refactor!: remove `wallet` and `rusqlite` features
The `bdk_chain` and `kyoto` crates can already interop with little developer overhead.
There is an additional burden to maintain the `wallet` feature, as opposed to making it the only approach to using the light client.
I am open to reconsidering wallet-less workflows if it is requested, but ultimately I think we should make an opinionated approach
and assume the user will have a `Wallet`. Docs and example burden goes down significantly and the workflow is clearer for users.