fix!: remove `LogSubscriber` and `WarningSubscriber`
I noticed after the example was fully synced to the network,
my computer started working overtime. I believe the loops involved
in these futures are not implemented properly, as in they should
actually register with the waker to be polled later. Instead they
are just busy looping aimlessly. Conceptually the `Receiver` type
from `tokio` is not that difficult to understand, so I return them
directly and inline the docs.