refactor(builder)!: prefer extension trait instead of wrapper
For every `N` methods on the `kyoto::NodeBuilder` there are `N`
things to add to this crate. Instead we can continue to abuse
the extension trait model and just tailor to the implementation
details of building a node for a wallet. Only unfortunate thing
here is the `ScanType` requires a bit of state with the wallet,
so it has to be passed as an argument.

Also remove deprecated `.clippy.toml`