refactor!: remove the optional `Update`
The optional case for `UpdateSubscriber::update` returning
a `None` value can only occur if the user:
1. Syncs the wallet to the latest chain tip
2. Closes the application
3. Reopens the application and re-syncs to the same tip

While this may happen there doesn't seem to be any harm in applying
the same update twice to a wallet. For developer ease-of-use, the
optional case can be ignored and a repeated update can be issued in
this edge case.