fix(client)!: return error if channel reports `None`
> This method returns None if the channel has been closed and there are
no remaining messages in the channel’s buffer. This indicates that no
further values can ever be received from this Receiver.

When a channel returns `None`, we know that all sending halves of the
channel have been dropped and no more messages may be received. In this
case that means that the node has stopped running, as there is only a
single sending halve that emits node events.