refactor: use `CheckPoint` for chain updates
From the docs:
> Checkpoints are cheaply cloneable

We can abuse that fact a bit and internally update a `CheckPoint`
whenever we receive chain events. With Kyoto able to emit what was
accepted to the new chain of most work, we should now be able to use
`CheckPoint::insert`, which seems to have the desired behavior.

1. If the checkpoint is already there, it is ignored
2. If there is a conflicting checkpoint, the older ones at or above the
   height will be removed, exactly what we want for a reorg
3. Anything inserted at a greater height is appended