Reveal SPKs iteratively
Bump `kyoto-cbf` to `0.15.0`, which has been re-released as `bip157`
`0.1.0`

The Kyoto API now emits filters as the only method of scanning the
chain. On the performance side, this allows the main loop within the
node to continue to download filters as they are being checked by the
underlying wallet. This should be a great improvement for users, as one
OS thread will be performing the expensive siphash24, and the other one
will be downloading filters and computing the CPU-optimized SHA256. On
scripts, the previous iteration of the crate could not recover a wallet
that has used more than `lookahead` number of scripts. Scripts are now
revealed interactively as they are found in blocks.