optimization: Add a task queue to fetch blocks
In the previous iteration, the interesting blocks are put in a queue and
fetched all at once when all filters have been emitted. The node can
still do work when fetching filters, namely fetch the blocks the wallet
needs. Here I introduce a queue of tasks that requests blocks in the
background. When at least one block is ready, it is applied to the
wallet. Once all filters have been emitted, the remaining tasks, if any,
are awaited and the blocks are applied. This should resemble previous
behavior when the node handled the scripts.