fix: Extend the `spk_cache`
First, as blocks are resolved from the new task queue, the
`peek_scripts` member function is called every time. This is wasteful in
compute, as users are already responsible for estimating the number of
used scripts when recovering a wallet. Secondly, the `peek_scripts`
returns a set, which was previously not being used. Now when the tip is
reached, actually add the scripts to the SPK cache.